# admin_openapi.yaml
openapi: 3.0.3
info:
  title: Admin BFF
  version: 1.0.0
  description: |
    API gateway for players, wallets, withdrawals, and reports services.
    Conventions:
      - Array fields are always arrays (empty array when no data).
      - Optional fields may be null when data is unavailable.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /finances/api/v1/withdrawal-requests:
    post:
      tags: [Withdrawals]
      summary: Create withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWithdrawalRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
  /finances/api/v1/withdrawal-requests/{id}:
    get:
      tags: [Withdrawals]
      summary: Get withdrawal request
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Withdrawal request
          content:
            application/json:
              schema:
                type: object
  /finances/api/v1/withdrawal-requests/{id}/take:
    post:
      tags: [Withdrawals]
      summary: Take withdrawal request
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/managerId"
        - $ref: "#/components/parameters/managerRole"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
  /finances/api/v1/withdrawal-requests/{id}/request-verification:
    post:
      tags: [Withdrawals]
      summary: Request verification
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/managerId"
        - $ref: "#/components/parameters/managerRole"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
  /finances/api/v1/withdrawal-requests/{id}/approve:
    post:
      tags: [Withdrawals]
      summary: Approve withdrawal
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/managerId"
        - $ref: "#/components/parameters/managerRole"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
  /finances/api/v1/withdrawal-requests/{id}/reject:
    post:
      tags: [Withdrawals]
      summary: Reject withdrawal
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/managerId"
        - $ref: "#/components/parameters/managerRole"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectWithdrawalRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
  /finances/api/v1/withdrawal-requests/{id}/retry:
    post:
      tags: [Withdrawals]
      summary: Retry withdrawal
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/managerId"
        - $ref: "#/components/parameters/managerRole"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
  /finances/api/v1/withdrawal-requests/{id}/return:
    post:
      tags: [Withdrawals]
      summary: Return withdrawal
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/managerId"
        - $ref: "#/components/parameters/managerRole"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
...
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    managerId:
      name: X-Manager-Id
      in: header
      required: true
      schema:
        type: string
    managerRole:
      name: X-Manager-Role
      in: header
      required: true
      schema:
        type: string
  schemas:
...
    CreateWithdrawalRequest:
      type: object
      properties:
        playerId:
          type: string
          format: uuid
        walletId:
          type: string
          format: uuid
        currency:
          type: string
        amount:
          type: string
        paymentSystemId:
          type: string
      required: [playerId, walletId, currency, amount, paymentSystemId]
    RejectWithdrawalRequest:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
...